+++ Obsidian/Blocks/BlockFloorSign.cs
@@ -2,7 +2,7 @@
 {
     public class BlockFloorSign : Block
     {
-        internal BlockFloorSign(string name, int id, Materials type) : base(name, id, type)
+        internal BlockFloorSign(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
