+++ Obsidian/Blocks/BlockFlower.cs
@@ -2,7 +2,7 @@
 {
     public class BlockFlower : Block
     {
-        internal BlockFlower(string name, int id, Materials type) : base(name, id, type)
+        internal BlockFlower(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
