+++ Obsidian/Blocks/BlockFluid.cs
@@ -2,7 +2,7 @@
 {
     public class BlockFluid : Block
     {
-        internal BlockFluid(string name, int id, Materials type) : base(name, id, type)
+        internal BlockFluid(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
