+++ Obsidian/Blocks/BlockGrass.cs
@@ -2,7 +2,7 @@
 {
     public class BlockGrass : Block
     {
-        internal BlockGrass(string name, int id, Materials type) : base(name, id, type)
+        internal BlockGrass(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
