+++ Obsidian/Blocks/BlockLeaves.cs
@@ -2,7 +2,7 @@
 {
     public class BlockLeaves : Block
     {
-        internal BlockLeaves(string name, int id, Materials type) : base(name, id, type)
+        internal BlockLeaves(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
