+++ Obsidian/Blocks/BlockLog.cs
@@ -2,7 +2,7 @@
 {
     public class BlockLog : Block
     {
-        internal BlockLog(string name, int id, Materials type) : base(name, id, type)
+        internal BlockLog(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
