+++ Obsidian/Blocks/BlockLongGrass.cs
@@ -2,7 +2,7 @@
 {
     public class BlockLongGrass : Block
     {
-        internal BlockLongGrass(string name, int id, Materials type) : base(name, id, type)
+        internal BlockLongGrass(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
