+++ Obsidian/Blocks/BlockNote.cs
@@ -2,7 +2,7 @@
 {
     public class BlockNote : Block
     {
-        internal BlockNote(string name, int id, Materials type) : base(name, id, type)
+        internal BlockNote(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
