+++ Obsidian/Blocks/BlockPoweredRail.cs
@@ -2,7 +2,7 @@
 {
     public class BlockPoweredRail : Block
     {
-        internal BlockPoweredRail(string name, int id, Materials type) : base(name, id, type)
+        internal BlockPoweredRail(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
