+++ Obsidian/Blocks/BlockReed.cs
@@ -2,7 +2,7 @@
 {
     public class BlockReed : Block
     {
-        internal BlockReed(string name, int id, Materials type) : base(name, id, type)
+        internal BlockReed(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
