+++ Obsidian/Blocks/BlockStainedGlass.cs
@@ -2,7 +2,7 @@
 {
     public class BlockStainedGlass : Block
     {
-        internal BlockStainedGlass(string name, int id, Materials type) : base(name, id, type)
+        internal BlockStainedGlass(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
