+++ Obsidian/Blocks/BlockWallTorch.cs
@@ -2,7 +2,7 @@
 {
     public class BlockWallTorch : BlockTorch
     {
-        internal BlockWallTorch(string name, int id) : base(name, id, Materials.WallTorch)
+        internal BlockWallTorch(string name, short id) : base(name, id, Materials.WallTorch)
         {
         }
     }
